feature: select2_filter
Allow filling in the filter without requiring that one option be selected.
This is useful for testing search filters are showing the right results.

Add a field option as an alternative way of finding a label by its "for" field.
This can be more stable than by the label text.